slight mods

+++ AlgorithmsLib/src/TurboXIC.cpp
@@ -141,7 +141,7 @@ XICPoints TurboXIC::Private::extractPointsXIC(
 
     for (const rTreePoint &rtp : result) {
         const auto scanNumber = static_cast<ScanNumber>(rtp.first.get<0>());
-        xicPoints.scanNumbersVsIntensityVals[scanNumber] += rtp.second;
+        xicPoints.scanNumbersVsIntensityVals[scanNumber] += static_cast<double>(rtp.second);
 //        xicPoints.scanNumberVsMzVals[scanNumber].push_back(rtp.first.get<1>());
         xicPoints.scanNumbersVsScanPoints[scanNumber].push_back({rtp.first.get<1>(), rtp.second});
     }
